git-svn-id: http://svn.code.sf.net/p/shopizer/code/shopizer-portlet@1633 e80ee2c7-76f3-4005-aeaf-98396743f455

+++ sm-shop/src/main/java/com/salesmanager/web/admin/controller/products/ProductController.java
@@ -730,7 +730,7 @@ public class ProductController {
 			
 			Long id = Long.parseLong(iid);
 			ProductImage productImage = productImageService.getById(id);
-			if(productImage==null || productImage.getProduct().getMerchantStore().getId()!=store.getId()) {
+			if(productImage==null || productImage.getProduct().getMerchantStore().getId().intValue()!=store.getId().intValue()) {
 
 				resp.setStatusMessage(messages.getMessage("message.unauthorized", locale));
 				resp.setStatus(AjaxResponse.RESPONSE_STATUS_FAIURE);
